Improve API
This collapses the user-external API down into just
`deploy_artifact_paths()` which takes in a `strategy` kwarg, which can
be one of `:hardlink`, `:copy` or `:symlink`.  In general, we auto-probe
to see if hardlinks are possible (they are not allowed if crossing a
device boundary) and if they are, we use that.  Otherwise, we prefer to
use `:copy`, as that's safest.  The user can request `:symlink` if they
know that it works for their usecase.