Attribute dependency artifacts to requested top-level JLLS
Previously, we would return a simple `source JLL => artifacts` mapping
as the output of `collect_artifact_paths()`, but this made it
challenging to group together dependencies properly.  Instead, we now
treat every JLL argument to `collect_artifact_paths()` as a top-level
request, resolve them together, then return the set of all recursive
dependencies of each top-level JLL in that dictionary.  This allows for
easy determination of which artifacts are used by which top-level
requests.  `flatten_artifact_paths()` now deduplicates its result.

Because this technically changes the return data layout, I opted to make
this a major version bump.