Ensure collecting already-installed deps doesn't modify Manifest.toml
This ensures that we never call `Pkg.add()` if the JLL we're interested
in already exists within the manifest.  If you place constraints on the
JLL such that the already-installed version is incompatible, we update
the manifest, but as long as it is satisfiable, the manifest should not
be touched.

Also improve tests to no longer install things into the main depot,
which makes it much less predictable.